Move parentTaskId into TransportRequest
Now everything can have a parent!
