+++ core/src/main/java/org/elasticsearch/transport/TransportRequest.java
@@ -19,17 +19,40 @@
 
 package org.elasticsearch.transport;
 
+import org.elasticsearch.common.io.stream.StreamInput;
+import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.tasks.Task;
+import org.elasticsearch.tasks.TaskId;
+
+import java.io.IOException;
 
 /**
  */
 public abstract class TransportRequest extends TransportMessage {
-
     public static class Empty extends TransportRequest {
         public static final Empty INSTANCE = new Empty();
     }
 
+    /**
+     * Parent of this request. Defaults to {@link TaskId#EMPTY_TASK_ID}, meaning "no parent".
+     */
+    private TaskId parentTaskId = TaskId.EMPTY_TASK_ID;
+
     public TransportRequest() {
+    }
+
+    /**
+     * Set a reference to task that caused this task to be run.
+     */
+    public void setParentTask(String parentTaskNode, long parentTaskId) {
+        setParentTask(new TaskId(parentTaskNode, parentTaskId));
+    }
+
+    /**
+     * Set a reference to task that caused this task to be run.
+     */
+    public void setParentTask(TaskId taskId) {
+        this.parentTaskId = taskId;
     }
 
     /**
@@ -41,10 +64,26 @@ public abstract class TransportRequest extends TransportMessage {
         return new Task(id, type, action, getDescription());
     }
 
+    public Task createTask(long id, String type, String action, TaskId parentTaskId) {
+        return new Task(id, type, action, getDescription(), parentTaskId);
+    }
+
     /**
      * Returns optional description of the request to be displayed by the task manager
      */
     public String getDescription() {
         return "";
+    }
+
+    @Override
+    public void readFrom(StreamInput in) throws IOException {
+        super.readFrom(in);
+        parentTaskId = TaskId.readFromStream(in);
+    }
+
+    @Override
+    public void writeTo(StreamOutput out) throws IOException {
+        super.writeTo(out);
+        parentTaskId.writeTo(out);
     }
 }
